Struct encoding_io::TextReader

pub struct TextReader<R> { /* fields omitted */ }

TextReader is a wrapper for a non-utf8 encoded text reader. It behaves as a utf8 encoded text reader.

Examples:

use std::fs::File;
use std::io::{Read, BufRead};
use encoding_io::TextReader;
use encoding_rs::*;
let mut reader = TextReader::new(
        File::open("shiftjis-text.txt").expect("failed to open."),
        SHIFT_JIS);
for line in reader.lines() {
    let line = line.expect("Failed to read.");
    println!("{}", line);
}

Methods

impl<R: Read> TextReader<R>

fn new(reader: R, encoding: &'static Encoding) -> TextReader<R>

Creates a new TextReader with a default buffer capacity.

fn with_capacity(
reader: R,
encoding: &'static Encoding,
unprocessed_buf_capacity: usize,
decoded_buf_capacity: usize
) -> TextReader<R>

Creates a new TextReader with the specified buffer capacity.

fn with_decoder(reader: R, decoder: Decoder) -> TextReader<R>

Creates a new TextReader with the specified decoder and a default buffer capacity.

fn with_decoder_and_capacity(
reader: R,
decoder: Decoder,
unprocessed_buf_capacity: usize,
decoded_buf_capacity: usize
) -> TextReader<R>

Creates a new TextReader with the specified decoder and specified buffer capacity.

fn into_decoder(self) -> Decoder

Consumes the TextReader and returns the underlying decoder.

fn into_inner(self) -> R

Consumes the TextReader and returns the underlying reader.

fn into_decoder_and_inner(self) -> (Decoder, R)

Consumes the TextReader and returns the underlying reader and decoder.