Struct encoding_io::TextWriter

pub struct TextWriter<W> { /* fields omitted */ }

TextWriter is a wrapper for writer. It encodes given bytes to utf8 then write to a writer.

Examples use std::fs::File; use std::io::Write; use encoding_io::TextWriter; use encoding_rs::*; let mut writer = TextWriter::new( File::create("shiftjis-text.txt").expect("Failed to create."), SHIFT_JIS); writer.write("some utf8 text".as_bytes()).expect("Failed to write.");

Methods

impl<W: Write> TextWriter<W>

fn new(writer: W, encoding: &'static Encoding) -> TextWriter<W>

Creates a new TextWriter with a default buffer capacity.

fn with_capacity(
writer: W,
encoding: &'static Encoding,
encoded_buf_capacity: usize
) -> TextWriter<W>

Creates a new TextWriter with the specified buffer capacity.

fn with_encoder(writer: W, encoder: Encoder) -> TextWriter<W>

Creates a new TextWriter with the specified encoder.

fn with_encoder_and_capacity(
writer: W,
encoder: Encoder,
encoded_buf_capacity: usize
) -> TextWriter<W>

Creates a new TextWriter with the specified encoder and specified buffer capacity.

fn into_encoder(self) -> Encoder

Consumes the TextWriter and returns the underlying encoder.

fn into_inner(self) -> W

Consumes the TextWriter and returns the underlying writer.

fn into_encoder_and_inner(self) -> (Encoder, W)

Consumes the TextWriter and returns the underlying writer and encoder.

fn finish(&mut self) -> Result<()>

Notice to encoder that the end of the stream is reached when all the characters in src have been consumed. This argument is needed for ISO-2022-JP and is ignored for other encodings.